code review findings fixed
