+++ src/app/_localization/README.md
@@ -1,7 +1,5 @@
 ## Translate contentful entries
 
-### 1. Wrap the page you want to translate in <LocalizedContentfulPage>
+Use `<LocalizedContentfulPageToggleButton>` to translate contentful entries with a click.
 
-### 2. Place one or more <LocalizedContentfulPageToggleButton> inside the page / as child of the page component
-
-### 3. Use one or multiple <LocalizedContentfulEntryField> inside the page component to translate the fields of contentful entries
+All `<LocalizedContentfulEntryField>` components will switch the locale when the button is clicked.
